requirements.txt
django==4.0.6
docker==5.0.3
dynaconf==3.1.9
numpy==1.23.1
pandas==1.4.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.1
requests==2.28.1
psycopg2-binary==2.9.3
redis==4.3.4
django-celery-beat==2.3.0
djangorestframework>=3.12.4
Faker==14.2.0
celery==5.2.7
flower==1.2.0